+++ ch.ahoegger.docbox.server/src/main/java/ch/ahoegger/docbox/server/document/DocumentPartnerService.java
@@ -77,7 +77,7 @@ public class DocumentPartnerService implements IDocumentPartnerTable {
 
   }
 
-  public void deletePartnerId(Long partnerId) {
+  public void deletePartnerId(BigDecimal partnerId) {
     StringBuilder statementBuilder = new StringBuilder();
     statementBuilder.append("DELETE FROM ").append(TABLE_NAME).append(" WHERE ").append(PARTNER_NR).append(" = :partnerId");
     SQL.delete(statementBuilder.toString(), new NVPair("partnerId", partnerId));
